[mage-1438] (#85)
1) Fix non-point observations not showing the associated form icon in the observations list
2) Refactor MapAnnotation's factory methods and associated style classes to make the logic clearer
