+++ mage/src/main/java/mil/nga/giat/mage/map/annotation/BaseObservationStyle.kt
@@ -1,0 +1,6 @@
+package mil.nga.giat.mage.map.annotation
+
+import android.net.Uri
+
+//for Point observation image icons
+open class BaseObservationStyle(val uri: Uri? = null)
